+++ traffic_portal/app/src/common/modules/table/deliveryServiceCapabilities/index.js
@@ -17,34 +17,5 @@
  * under the License.
  */
 
-exports.config = {
-	framework: 'jasmine',
-	seleniumAddress: 'http://localhost:4444/wd/hub',
-	baseUrl: 'https://localhost:4443',
-	getPageTimeout: 30000,
-
-	capabilities: {
-		'browserName': 'chrome'
-	},
-	params: {
-		adminUser: 'admin',
-		adminPassword: 'twelve'
-	},
-	jasmineNodeOpts: {defaultTimeoutInterval: 600000},
-
-	suites: {
-		loginTests: 'login/login-spec.js',
-		allTests: [
-			'login/login-spec.js',
-			'CDNs/cdns-spec.js',
-			'cacheGroups/cache-groups-spec.js',
-			'profiles/profiles-spec.js',
-			'divisions/divisions-spec.js',
-			'regions/regions-spec.js',
-			'physLocations/phys-locations-spec.js',
-			'serverCapabilities/server-capabilities-spec.js',
-			'deliveryServices/delivery-services-spec.js',
-			'servers/servers-spec.js'
-		]
-	}
-};
+module.exports = angular.module('trafficPortal.table.deliveryServiceCapabilities', [])
+	.controller('TableDeliveryServiceCapabilitiesController', require('./TableDeliveryServiceCapabilitiesController'));